Move SSO component into its own folder for better folder management for future components in auth.
