+++ bitwarden_license/bit-web/src/app/auth/sso/sso.component.ts
